reorganised graphql configuration classes
