+++ src/main/java/li/naska/bgg/graphql/controller/GuildController.java
@@ -7,18 +7,35 @@ import li.naska.bgg.graphql.model.Address;
 import li.naska.bgg.graphql.model.Guild;
 import li.naska.bgg.graphql.model.GuildMember;
 import li.naska.bgg.graphql.model.User;
+import li.naska.bgg.graphql.service.GraphQLGuildsService;
 import org.dataloader.DataLoader;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
+import org.springframework.graphql.execution.BatchLoaderRegistry;
 import org.springframework.stereotype.Controller;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import reactor.util.function.Tuple2;
 
 import java.time.ZonedDateTime;
 import java.util.List;
 
 @Controller("GraphQLGuildController")
 public class GuildController {
+
+  public GuildController(BatchLoaderRegistry registry, GraphQLGuildsService guildsService) {
+    registry.forTypePair(Integer.class, GuildV2.class).registerMappedBatchLoader((ids, env) ->
+        Flux.fromIterable(ids)
+            .flatMap(id -> Mono.just(id).zipWith(guildsService.getGuild(id)))
+            .collectMap(Tuple2::getT1, tuple -> new GuildV2(tuple.getT2()))
+    );
+    registry.forTypePair(Integer.class, GuildV2Members.class).registerMappedBatchLoader((ids, env) ->
+        Flux.fromIterable(ids)
+            .flatMap(id -> Mono.just(id).zipWith(guildsService.getMembers(id)))
+            .collectMap(Tuple2::getT1, tuple -> new GuildV2Members(tuple.getT2()))
+    );
+  }
 
   @QueryMapping
   public Mono<Guild> guildById(@Argument Integer id, DataLoader<Integer, GuildV2> loader) {
