+++ src/main/java/li/naska/bgg/graphql/controller/RankingController.java
@@ -3,16 +3,29 @@ package li.naska.bgg.graphql.controller;
 import li.naska.bgg.graphql.data.UserV2Ranking;
 import li.naska.bgg.graphql.model.RankedItem;
 import li.naska.bgg.graphql.model.Ranking;
+import li.naska.bgg.graphql.model.enums.Domain;
+import li.naska.bgg.graphql.service.GraphQLUsersService;
 import org.dataloader.DataLoader;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
+import org.springframework.graphql.execution.BatchLoaderRegistry;
 import org.springframework.stereotype.Controller;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import reactor.util.function.Tuple2;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Controller("GraphQLRankingController")
 public class RankingController {
+
+  public RankingController(BatchLoaderRegistry registry, GraphQLUsersService usersService) {
+    registry.forTypePair(UserV2Ranking.UserRankingKey.class, UserV2Ranking.class).registerMappedBatchLoader((keys, env) ->
+        Flux.fromIterable(keys)
+            .flatMap(key -> Mono.just(key).zipWith(usersService.getUserRanking(key.username(), key.type(), Domain.valueOf(key.domain()))))
+            .collectMap(Tuple2::getT1, tuple -> new UserV2Ranking(tuple.getT2()))
+    );
+  }
 
   @SchemaMapping
   public Mono<List<RankedItem>> boardgame(Ranking ranking, DataLoader<UserV2Ranking.UserRankingKey, UserV2Ranking> loader) {
